pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lvgpt"
version = "0.1.0"
description = "LabVIEW graph builder for LLM-generated code"
authors = [{ name = "LVGPT" }]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["pytest", "jsonschema", "pyyaml"]

[project.scripts]
lvgpt-build = "lvgpt.cli:main"
lvgpt-build-ir = "lvgpt.cli_ir:main"

[tool.setuptools]
package-dir = {"" = "python"}
packages = ["lvgpt", "lvgpt.lvgraph"]
